RF: Make sure MicrophoneDevice exists in psychopy.sound for legacy
