+++ psychopy/hardware/microphone.py
@@ -1,16 +1,28 @@
 import sys
 import time
 
+import numpy as np
 from psychtoolbox import audio as audio
-
 from psychopy import logging as logging, prefs
 from psychopy.constants import NOT_STARTED
 from psychopy.hardware import BaseDevice
 from psychopy.sound import AudioDeviceInfo, AudioDeviceStatus, AudioClip
 from psychopy.sound.exceptions import AudioInvalidCaptureDeviceError, AudioInvalidDeviceError, \
-    AudioStreamError
-from psychopy.sound.microphone import _hasPTB, RecordingBuffer
+    AudioStreamError, AudioRecordingBufferFullError
 from psychopy.tools import systemtools as st
+from psychopy.tools.audiotools import SAMPLE_RATE_48kHz
+
+
+_hasPTB = True
+try:
+    import psychtoolbox.audio as audio
+except (ImportError, ModuleNotFoundError):
+    logging.warning(
+        "The 'psychtoolbox' library cannot be loaded but is required for audio "
+        "capture (use `pip install psychtoolbox` to get it). Microphone "
+        "recording will be unavailable this session. Note that opening a "
+        "microphone stream will raise an error.")
+    _hasPTB = False
 
 
 class MicrophoneDevice(BaseDevice, aliases=["mic", "microphone"]):
@@ -697,3 +709,279 @@ class MicrophoneDevice(BaseDevice, aliases=["mic", "microphone"]):
                 "call `Microphone.stop` first.")
 
         return self._recording.getSegment()  # full recording
+
+
+class RecordingBuffer:
+    """Class for a storing a recording from a stream.
+
+    Think of instances of this class behaving like an audio tape whereas the
+    `MicrophoneDevice` class is the tape recorder. Samples taken from the stream are
+    written to the tape which stores the data.
+
+    Used internally by the `MicrophoneDevice` class, users usually do not create
+    instances of this class themselves.
+
+    Parameters
+    ----------
+    sampleRateHz : int
+        Sampling rate for audio recording in Hertz (Hz). By default, 48kHz
+        (``sampleRateHz=48000``) is used which is adequate for most consumer
+        grade microphones (headsets and built-in).
+    channels : int
+        Number of channels to record samples to `1=Mono` and `2=Stereo`.
+    maxRecordingSize : int
+        Maximum recording size in kilobytes (Kb). Since audio recordings tend to
+        consume a large amount of system memory, one might want to limit the
+        size of the recording buffer to ensure that the application does not run
+        out of memory. By default, the recording buffer is set to 24000 KB (or
+        24 MB). At a sample rate of 48kHz, this will result in 62.5 seconds of
+        continuous audio being recorded before the buffer is full.
+    policyWhenFull : str
+        What to do when the recording buffer is full and cannot accept any more
+        samples. If 'ignore', samples will be silently dropped and the `isFull`
+        property will be set to `True`. If 'warn', a warning will be logged and
+        the `isFull` flag will be set. Finally, if 'error' the application will
+        raise an exception.
+
+    """
+    def __init__(self, sampleRateHz=SAMPLE_RATE_48kHz, channels=2,
+                 maxRecordingSize=24000, policyWhenFull='ignore'):
+        self._channels = channels
+        self._sampleRateHz = sampleRateHz
+        self._maxRecordingSize = maxRecordingSize
+        self._samples = None  # `ndarray` created in _allocRecBuffer`
+        self._offset = 0  # recording offset
+        self._lastSample = 0  # offset of the last sample from stream
+        self._spaceRemaining = None  # set in `_allocRecBuffer`
+        self._totalSamples = None  # set in `_allocRecBuffer`
+
+        # check if the value is valid
+        if policyWhenFull not in ['ignore', 'warn', 'error']:
+            raise ValueError("Invalid value for `policyWhenFull`.")
+
+        self._policyWhenFull = policyWhenFull
+        self._warnedRecBufferFull = False
+        self._loops = 0
+
+        self._allocRecBuffer()
+
+    def _allocRecBuffer(self):
+        """Allocate the recording buffer. Called internally if properties are
+        changed."""
+        # allocate another array
+        nBytes = self._maxRecordingSize * 1000
+        recArraySize = int((nBytes / self._channels) / (np.float32()).itemsize)
+
+        self._samples = np.zeros(
+            (recArraySize, self._channels), dtype=np.float32, order='C')
+
+        # sanity check
+        assert self._samples.nbytes == nBytes
+        self._totalSamples = len(self._samples)
+        self._spaceRemaining = self._totalSamples
+
+    @property
+    def samples(self):
+        """Reference to the actual sample buffer (`ndarray`)."""
+        return self._samples
+
+    @property
+    def bufferSecs(self):
+        """Capacity of the recording buffer in seconds (`float`)."""
+        return self._totalSamples / self._sampleRateHz
+
+    @property
+    def nbytes(self):
+        """Number of bytes the recording buffer occupies in memory (`int`)."""
+        return self._samples.nbytes
+
+    @property
+    def sampleBytes(self):
+        """Number of bytes per sample (`int`)."""
+        return np.float32().itemsize
+
+    @property
+    def spaceRemaining(self):
+        """The space remaining in the recording buffer (`int`). Indicates the
+        number of samples that the buffer can still add before overflowing.
+        """
+        return self._spaceRemaining
+
+    @property
+    def isFull(self):
+        """Is the recording buffer full (`bool`)."""
+        return self._spaceRemaining <= 0
+
+    @property
+    def totalSamples(self):
+        """Total number samples the recording buffer can hold (`int`)."""
+        return self._totalSamples
+
+    @property
+    def writeOffset(self):
+        """Index in the sample buffer where new samples will be written when
+        `write()` is called (`int`).
+        """
+        return self._offset
+
+    @property
+    def lastSample(self):
+        """Index of the last sample recorded (`int`). This can be used to slice
+        the recording buffer, only getting data from the beginning to place
+        where the last sample was written to.
+        """
+        return self._lastSample
+
+    @property
+    def loopCount(self):
+        """Number of times the recording buffer restarted (`int`). Only valid if
+        `loopback` is ``True``."""
+        return self._loops
+
+    @property
+    def maxRecordingSize(self):
+        """Maximum recording size in kilobytes (`int`).
+
+        Since audio recordings tend to consume a large amount of system memory,
+        one might want to limit the size of the recording buffer to ensure that
+        the application does not run out of memory. By default, the recording
+        buffer is set to 24000 KB (or 24 MB). At a sample rate of 48kHz, this
+        will result in 62.5 seconds of continuous audio being recorded before
+        the buffer is full.
+
+        Setting this value will allocate another recording buffer of appropriate
+        size. Avoid doing this in any time sensitive parts of your application.
+
+        """
+        return self._maxRecordingSize
+
+    @maxRecordingSize.setter
+    def maxRecordingSize(self, value):
+        value = int(value)
+
+        # don't do this unless the value changed
+        if value == self._maxRecordingSize:
+            return
+
+        # if different than last value, update the recording buffer
+        self._maxRecordingSize = value
+        self._allocRecBuffer()
+
+    def seek(self, offset, absolute=False):
+        """Set the write offset.
+
+        Use this to specify where to begin writing samples the next time `write`
+        is called. You should call `seek(0)` when starting a new recording.
+
+        Parameters
+        ----------
+        offset : int
+            Position in the sample buffer to set.
+        absolute : bool
+            Use absolute positioning. Use relative positioning if `False` where
+            the value of `offset` will be added to the current offset. Default
+            is `False`.
+
+        """
+        if not absolute:
+            self._offset += offset
+        else:
+            self._offset = absolute
+
+        assert 0 <= self._offset < self._totalSamples
+        self._spaceRemaining = self._totalSamples - self._offset
+
+    def write(self, samples):
+        """Write samples to the recording buffer.
+
+        Parameters
+        ----------
+        samples : ArrayLike
+            Samples to write to the recording buffer, usually of a stream. Must
+            have the same number of dimensions as the internal array.
+
+        Returns
+        -------
+        int
+            Number of samples overflowed. If this is zero then all samples have
+            been recorded, if not, the number of samples rejected is given.
+
+        """
+        nSamples = len(samples)
+        if self.isFull:
+            if self._policyWhenFull == 'ignore':
+                return nSamples  # samples lost
+            elif self._policyWhenFull == 'warn':
+                if not self._warnedRecBufferFull:
+                    logging.warning(
+                        f"Audio recording buffer filled! This means that no "
+                        f"samples are saved beyond {round(self.bufferSecs, 6)} "
+                        f"seconds. Specify a larger recording buffer next time "
+                        f"to avoid data loss.")
+                    logging.flush()
+                    self._warnedRecBufferFull = True
+                return nSamples
+            elif self._policyWhenFull == 'error':
+                raise AudioRecordingBufferFullError(
+                    "Cannot write samples, recording buffer is full.")
+            else:
+                return nSamples  # whatever
+
+        if not nSamples:  # no samples came out of the stream, just return
+            return
+
+        if self._spaceRemaining >= nSamples:
+            self._lastSample = self._offset + nSamples
+            audioData = samples[:, :]
+        else:
+            self._lastSample = self._offset + self._spaceRemaining
+            audioData = samples[:self._spaceRemaining, :]
+
+        self._samples[self._offset:self._lastSample, :] = audioData
+        self._offset += nSamples
+
+        self._spaceRemaining -= nSamples
+
+        # Check if the recording buffer is now full. Next call to `poll` will
+        # not record anything.
+        if self._spaceRemaining <= 0:
+            self._spaceRemaining = 0
+
+        d = nSamples - self._spaceRemaining
+        return 0 if d < 0 else d
+
+    def clear(self):
+        # reset all live attributes
+        self._samples = None
+        self._offset = 0
+        self._lastSample = 0
+        self._spaceRemaining = None
+        self._totalSamples = None
+        # reallocate buffer
+        self._allocRecBuffer()
+
+    def getSegment(self, start=0, end=None):
+        """Get a segment of recording data as an `AudioClip`.
+
+        Parameters
+        ----------
+        start : float or int
+            Absolute time in seconds for the start of the clip.
+        end : float or int
+            Absolute time in seconds for the end of the clip. If `None` the time
+            at the last sample is used.
+
+        Returns
+        -------
+        AudioClip
+            Audio clip object with samples between `start` and `end`.
+
+        """
+        idxStart = int(start * self._sampleRateHz)
+        idxEnd = self._lastSample if end is None else int(
+            end * self._sampleRateHz)
+
+        return AudioClip(
+            np.array(self._samples[idxStart:idxEnd, :],
+                     dtype=np.float32, order='C'),
+            sampleRateHz=self._sampleRateHz)
